把auth/routes.js合併到app.js裡,將passport的name寫到session裡
