+++ controllers/users/logout.js
@@ -1,4 +1,5 @@
 module.exports = function (req, res){
   req.session.user = null;
+  req.logout();
   res.redirect('/');
 };
