+++ controllers/users/overview.js
@@ -1,7 +1,7 @@
 var models = require('../../models');
 
 module.exports = function (req, res){
-  models.posts.find({name: req.user.google.name}).sort({create_date:-1}).exec(function(err, posts){
+  models.posts.find({name: req.session.user}).sort({create_date:-1}).exec(function(err, posts){
     if(err){
       console.error(err);
       return res.send(err);
